+++ app/models/disco_app/concerns/shop.rb
@@ -61,6 +61,11 @@ module DiscoApp::Concerns::Shop
       "https://#{shopify_domain}/admin"
     end
 
+    # Convenience method to get the email of the shop's admin, to display in Rollbar.
+    def email
+      self.data['email']
+    end
+
     def installed_duration
       distance_of_time_in_words_to_now(created_at.time)
     end